[v0.89.1][maintainability] Split execute_tests/delegation_resume.rs by responsibility (Closes #1946)
